src/player/requirements.txt
flask 
requests
psycopg2-binary